% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtleviewer2.R
\docType{package}
\name{turtleviewer2-package}
\alias{turtleviewer2}
\alias{turtleviewer2-package}
\title{turtleviewer2: Turtle Data Viewer}
\description{
\code{turtleviewer2} is an interactive dashboard for
Western Australian turtle monitoring data,
hosted at \url{https://turtledata.dbca.wa.gov.au/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/turtleviewer2}
  \item Report bugs at \url{https://github.com/dbca-wa/turtleviewer2/issues}
}

}
\author{
\strong{Maintainer}: Florian W. Mayer \email{Florian.Mayer@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-4269-4242}{ORCID})

Other contributors:
\itemize{
  \item DBCA [copyright holder, funder]
  \item NWSFTCP [funder]
}

}
\keyword{internal}
